+++ src/index.js
@@ -1,11 +1,11 @@
 const { OrchestrationMachine, WorkflowError } = require('./orchestration-machine')
 const { MemoryRepository } = require('./memory-repository')
-const { HybirdPipeline, PipelineError } = require('./hybird-pipeline')
+const { HybridPipeline, PipelineError } = require('./hybrid-pipeline')
 
 module.exports = {
   MemoryRepository,
   OrchestrationMachine,
   WorkflowError,
-  HybirdPipeline,
+  HybridPipeline,
   PipelineError
 }